Split the title into album and artist
This is not ideal as we have to rely on the " - " delimiter.
There is no guarantee that it will be there, or that there won't
be more in the artist and album.

For now we remove the post if it does not contain " - " and we also
assume that the first " - " is the delimiter.