Handle extra cases from the post dump files
Sometimes the timestamps we get are strings, so we need to convert these
to integers.

A post could have also been deleted and we don't want to include these,
so we filter those out.

The dump files aren't wrapped in a "data" key like the api repsonses,
and we don't really need to store that so we get rid of it as well.