Deploy to fly.io rather than gigalixir
Remove a bunch of gigalixir specific stuff (buildpacks, package.json)
We also remove connecting to postgres via ssl as this is not necessary on fly.

Fly uses mix releases for deployment, which means we have to move some config into runtime.exs.

The rest of the files were auto-generated by fly launch.