Update post_to_release to remove invalid urls from thumbnail_url and url
thumbnail_url can return values like default, self, nsfw depending
on the post type. We may as well just check the url is valid to cover all bases.

Based on the posts we've imported already, I think the url is always going to be valid.
However there's no harm in checking this field as well.