Add period_header component
We extract the period header component from index.heex and also refactor
it to use the start and end date rather than indexes.

This also helps simplify the calculation for checking if the date is in the
latest period, so we can also remove the helper function from the Time module.

Also adjust the top_releases function to return the period start and
end dates as dates rather than datetimes as we don't need the extra info.