Add a more complex schedule for importing new releases.
We import every 30 mins on Thursday through to Saturday.
Most new music is released late Thursday / early Friday UTC time
and we also want to keep the data fresh on Saturday when the reddit
data will change the most.

The rest of the week we import every 2 hours as there is much less action.