Update title_to_artist_and_album to handle more cases
This now allows any type of unicode dash to be used as a delimiter and also handles titles with multiple delimiters in them.
We make the assumption that the rest of the delimiters are part of the album name as this is the most likely scenario.
There will be a few instances when this parsing doesn't get the correct result, but its better than them hanging around as in review.